refactoring of ports
